email_api/email_api/otp/serializers.py
from rest_framework import serializers
from .models import OTP


class OTPSerializer(serializers.ModelSerializer):
    class Meta:
        model = OTP
        fields = ['email', 'code', 'created_at', 'is_verified']
