email_api/requirements.txt
Django>=5.1.6
djangorestframework>=3.14.0
